pyproject.toml
[project]
name = "warped-diffusion"
version = "0.1.0"
description = "Semantic Differentiable Warping for Diffusion Models"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "diffusers>=0.25.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "peft>=0.7.0",
    "datasets>=2.16.0",
    "tensorboard>=2.15.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.0",
    "safetensors>=0.4.0",
    "lpips>=0.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/warped_diffusion"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
